# openapi.yaml
openapi: 3.0.3
info:
  title: Decibel Tribe API
  description: API specification for the Decibel Tribe application.
  version: 0.1.0
servers:
  - url: /api
    description: Local development server
paths:
...
  /events:
    get:
      summary: Get events (user or public)
      description: Retrieves a list of events. If the 'user' query parameter (username) is provided, it fetches events related to that user (respecting calendar privacy). Otherwise, it fetches events related to the authenticated user (created or attending published events). Requires authentication if no username is provided or to check private calendar.
      tags:
        - Events
      parameters:
        - name: user
          in: query
          description: Optional username to filter events for a specific user.
          required: false
          schema:
            type: string
      security:
        - cookieAuth: [] # Required for own events or checking other users' privacy
      responses:
        "200":
          description: A list of events matching the criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventData"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User specified in the query parameter not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: User not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a new event
      description: Creates a new event record associated with the authenticated user.
      tags:
        - Events
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventInput" # Use a specific input schema
      responses:
        "201":
          description: Event created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventData"
        "400":
          description: Invalid input data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: A message describing the error.
      required:
        - error
    UserSummary: # Basic user info used for embedding (based on getUserDataSelect)
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: url
          nullable: true
        createdAt: # Added from getUserDataSelect
          type: string
          format: date-time
        isAdmin: # Added from getUserDataSelect
          type: boolean
        _count: # Added from getUserDataSelect
          type: object
          properties:
            posts:
              type: integer
              description: Count of posts by this user.
            followers:
              type: integer
              description: Count of users following this user.
            following:
              type: integer
              description: Count of users this user is following.
          required: [posts, followers, following]
      required:
        - id
        - username
        - createdAt
        - isAdmin
        - _count
...
    EventAttendee: # Basic attendee record
      type: object
      properties:
        userId:
          type: string
        eventId:
          type: string
      required: [userId, eventId]
    EventAttendeeWithUser: # Attendee record including UserSummary
       allOf:
        - $ref: "#/components/schemas/EventAttendee"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/UserSummary"
          required: [user]
    EventData: # Represents event details (based on getEventDataInclude)
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
        url:
          type: string
          format: url
          nullable: true
        when:
          type: string
          format: date-time
        startTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
          example: "14:30"
        endTime:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
          example: "16:00"
        performers:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED] # Based on Prisma schema/usage
        visibility:
          type: string
          enum: [PRIVATE, PUBLIC] # Based on Prisma schema/usage
        isCancelled:
          type: boolean
        createdById:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy: # Embedded user data (owner)
          $ref: "#/components/schemas/UserSummary" # Uses getUserDataSelect
        attendees: # List of attendees with user details
          type: array
          items:
            $ref: "#/components/schemas/EventAttendeeWithUser" # Uses getUserDataSelect for user
        _count: # Count of attendees
          type: object
          properties:
            attendees:
              type: integer
          required: [attendees]
      required:
        - id
        - title
        - location
        - when
        - startTime
        - endTime
        - performers
        - status
        - visibility
        - isCancelled
        - createdById
        - createdAt
        - updatedAt
        - createdBy
        - attendees
        - _count
    CreateEventInput: # Schema for POST /events request body
       type: object
       properties:
         title:
           type: string
         location:
           type: string
         description:
           type: string
           nullable: true
         url:
           type: string
           format: url
           nullable: true
         when:
           type: string
           format: date-time
         startTime:
           type: string
           pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
         endTime:
           type: string
           pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$' # HH:mm format
         performers:
           type: array
           items:
             type: string
         status:
           type: string
           enum: [DRAFT, PUBLISHED]
         visibility:
           type: string
           enum: [PRIVATE, PUBLIC]
       required:
         - title
         - location
         - when
         - startTime
         - endTime
         - performers
         - status
         - visibility
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_session # Use the actual session cookie name from lucia config
      description: "Authentication cookie obtained via login or OAuth callback. Send requests with `credentials: 'include'` on the client."
    cronAuth: # New scheme for cron job Bearer token
      type: http
      scheme: bearer
      bearerFormat: plain # Assuming the secret is used directly
      description: Bearer token authentication required for cron job endpoints (using CRON_SECRET environment variable value).

  # ---------- Standard Responses ----------
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid (e.g., no session cookie, invalid session).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
...
    InternalServerError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Internal server error